HYD-4565 Update through module with bufferString and toJson
In an effort to utilize re-usable code in highland we created a
"through" module that will allow us to re-use common "through" blocks.
In this patch we are adding a couple of extra "through" components:

- buffer-lines
- buffer-string
- collect-values
- to-json

This patch will utilize dirp to simplify loading the components in the
index file. The above mentioned files will be added and will be loadable
through the index file. Additionally, i've setup the module for unit testing
and added test files for each associated source file.

- Added buffer-lines
- Added buffer-string
- Added collect-values
- Updated index to use dirp instead of adding each file individually
- Added jasmine setup
- Added limit.js
- Added to-json
- Added tests to all files

Signed-off-by: Will Johnson <william.c.johnson@intel.com>